Break out AVIOContext stuff into their own header and source files
