+++ src/torchcodec/decoders/_core/VideoDecoderOps.cpp
@@ -11,6 +11,7 @@
 #include <string>
 #include "c10/core/SymIntArrayRef.h"
 #include "c10/util/Exception.h"
+#include "src/torchcodec/decoders/_core/AVIOBytesContext.h"
 #include "src/torchcodec/decoders/_core/VideoDecoder.h"
 
 namespace facebook::torchcodec {
@@ -64,81 +65,6 @@ TORCH_LIBRARY(torchcodec_ns, m) {
 }
 
 namespace {
-
-// TODO: make comment below better
-// A struct that holds state for reading bytes from an IO context.
-// We give this to FFMPEG and it will pass it back to us when it needs to read
-// or seek in the memory buffer.
-//
-// A class that can be used as AVFormatContext's IO context. It reads from a
-// memory buffer that is passed in.
-class AVIOBytesContext : public AVIOContextHolder {
- public:
-  explicit AVIOBytesContext(const void* data, int64_t dataSize)
-      : dataContext_{static_cast<const uint8_t*>(data), dataSize, 0} {
-    TORCH_CHECK(data != nullptr, "Video data buffer cannot be nullptr!");
-    TORCH_CHECK(dataSize > 0, "Video data size must be positive");
-    createAVIOContext(&read, &seek, &dataContext_);
-  }
-
-  // The signature of this function is defined by FFMPEG.
-  static int read(void* opaque, uint8_t* buf, int buf_size) {
-    auto dataContext = static_cast<DataContext*>(opaque);
-    TORCH_CHECK(
-        dataContext->current <= dataContext->size,
-        "Tried to read outside of the buffer: current=",
-        dataContext->current,
-        ", size=",
-        dataContext->size);
-
-    buf_size = FFMIN(
-        buf_size, static_cast<int>(dataContext->size - dataContext->current));
-    TORCH_CHECK(
-        buf_size >= 0,
-        "Tried to read negative bytes: buf_size=",
-        buf_size,
-        ", size=",
-        dataContext->size,
-        ", current=",
-        dataContext->current);
-
-    if (!buf_size) {
-      return AVERROR_EOF;
-    }
-    memcpy(buf, dataContext->data + dataContext->current, buf_size);
-    dataContext->current += buf_size;
-    return buf_size;
-  }
-
-  // The signature of this function is defined by FFMPEG.
-  static int64_t seek(void* opaque, int64_t offset, int whence) {
-    auto dataContext = static_cast<DataContext*>(opaque);
-    int64_t ret = -1;
-
-    switch (whence) {
-      case AVSEEK_SIZE:
-        ret = dataContext->size;
-        break;
-      case SEEK_SET:
-        dataContext->current = offset;
-        ret = offset;
-        break;
-      default:
-        break;
-    }
-
-    return ret;
-  }
-
- private:
-  struct DataContext {
-    const uint8_t* data;
-    int64_t size;
-    int64_t current;
-  };
-
-  DataContext dataContext_;
-};
 
 at::Tensor wrapDecoderPointerToTensor(
     std::unique_ptr<VideoDecoder> uniqueDecoder) {
